+++ src/main/adapters/ffmpeg_adapter.py
@@ -1,4 +1,4 @@
-from core.ports.ffmpeg_port import FFmpegPort
+from main.core.ports.ffmpeg_port import FFmpegPort
 import os
 import subprocess
